fix: Prevent discarding editor WebView state during orientation change (#21662)
* fix: Prevent discarding editor WebView state during orientation change

The GutenbergKit editor relies upon a WebView, which discarded all its
state--including content edits--when the device orientation changed.
`EditPostActivity` is currently also utilized by Aztec and Gutenberg
Mobile, so there could be negative side effects to those fragments with
this change.

* fix: Experimental features activity relies upon OS-applied configuration

There is no explicit need for this configuration, it was mistakenly
copied during the initial creation.

* docs: Add release note

* fix: Increase robustness of managing editor configuration changes

Handle additional configurations and invalidate the editor WebView to
follow best practices.

https://developer.android.com/quick-guides/content/manage-webview-state