Merge release/25.7 into trunk (#21692)
* Download store metadata when generating beta build (#21623)

* Download store metadata when generating beta build

This should provide earlier feedback on the metadata translation, at
least for the portion that has already been translated at the time a
beta is made.

This is consistent with what we do in iOS, see
https://github.com/wordpress-mobile/WordPress-iOS/blob/b9a0deaf091739f33051f20cffdbe9a2097f645e/fastlane/lanes/release.rb#L219

* Remove call to push changes in download translated metadata lane

Automatically pushing from lanes other than the root-release lanes (code
freeze, new beta, finalize release, etc.) can result in undesired
pushes, like it just happened to me in
https://github.com/wordpress-mobile/WordPress-Android/commit/edcd40a47f4e4b80b049dc26f97aa335d7043659

* fix: Address editor crashes from null listeners (#21657)

* fix: Address editor crashes from null listeners

Setting the listeners within `instantiateItem` appears problematic as
this only runs for the initial launch, but not for other events that
destroy and recreate fragment--e.g., screen rotation events.

To address this, GutenbergKit now follow the pattern of the Aztec and
GutenbergMobile editors, invoking the `initializeEditorFragment` method.
This method is described as "requesting dependency injection," which
feels like an odd statement. There is likely further opportunity to
refactor this architecture for simplicity.

* docs: Add release note

* docs: Remove `TODO` comment

Apparently the linter prohibits this statement...

* Merge release_notes/25.7 into release/25.7 (#21691)

* Update 25.7 release notes

* Update WordPress `PlayStoreStrings.po` for version 25.7

* Update Jetpack `PlayStoreStrings.po` for version 25.7

---------

Co-authored-by: Oguz Kocer <oguz.kocer@automattic.com>

* Update translations

* Bump version number

* Update WordPress metadata translations for 25.7

* Update Jetpack metadata translations for 25.7

---------

Co-authored-by: Gio Lodi <gio.lodi@automattic.com>
Co-authored-by: David Calhoun <github@davidcalhoun.me>
Co-authored-by: Oguz Kocer <oguz.kocer@automattic.com>